requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-redis==5.4.0
django-rq==3.0.1
djangorestframework==3.16.0
Faker==37.1.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
redis==5.2.1
rq==2.3.2
sqlparse==0.5.3
tzdata==2025.2
